static node graphics to individual components
