+++ packages/nodes/src/components/nodes/generic-node/components/index.ts
@@ -1,3 +1,7 @@
 import GenericNodePort from './GenericNodePort'
 
 export { GenericNodePort }
+
+export { GenericNodeBody } from './GenericNodeBody'
+export { GenericNodeLabel } from './GenericNodeLabel'
+export { GenericNodeShadow } from './GenericNodeShadow'
